+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## [0.4.0] - 2023-12-27
+
 ### Added
 
 - Identifiers for BAS Data Catalogue, DOI and Esri